+++ crates/smtp/src/scripts/exec.rs
@@ -45,10 +45,7 @@ impl<T: AsyncWrite + AsyncRead + Unpin + IsTls> Session<T> {
             .set_variable("remote_ip", self.data.remote_ip.to_string())
             .set_variable("remote_ip.reverse", self.data.remote_ip.to_reverse_name())
             .set_variable("helo_domain", self.data.helo_domain.to_lowercase())
-            .set_variable(
-                "authenticated_as",
-                self.data.authenticated_as.to_lowercase(),
-            )
+            .set_variable("authenticated_as", self.data.authenticated_as.clone())
             .set_variable(
                 "now",
                 SystemTime::now()