+++ tests/src/smtp/mod.rs
@@ -238,6 +238,7 @@ impl TestConfig for SessionConfig {
                 errors_max: IfBlock::new(10),
                 errors_wait: IfBlock::new(Duration::from_secs(1)),
                 allow_plain_text: IfBlock::new(false),
+                must_match_sender: IfBlock::new(false),
             },
             mail: Mail {
                 script: IfBlock::new(None),